@@ -59,7 +59,7 @@ write_to = "src/multiplex_pipeline/_version.py"
 
 [tool.black]
 line-length = 79
-target-version = ['py38', 'py39', 'py310']
+target-version = ['py39', 'py310', 'py311']
 
 [tool.ruff]
 line-length = 79
